Move `Box` hooks
