+++ packages/fuselage/src/components/index.ts
@@ -4,7 +4,8 @@ export * from './AutoComplete';
 export * from './Avatar';
 export * from './Badge';
 export { default as Banner } from './Banner';
-export { Box, useArrayLikeClassNameProp } from './Box';
+export { Box } from './Box';
+export { useArrayLikeClassNameProp } from '../hooks/useArrayLikeClassNameProp';
 export { default as Button, ActionButton } from './Button';
 export * from './ButtonGroup';
 export * from './Callout';
